+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Add .gitattributes file
+
 ## [1.0.5] - 2025-06-14
 
 - Update Dart SDK to 3.8.0
@@ -19,6 +25,7 @@
 
 - Increase version
 
+[Unreleased]: https://github.com/ggsuite/gg_golden/compare/1.0.5...HEAD
 [1.0.5]: https://github.com/ggsuite/gg_golden/compare/1.0.4...1.0.5
 [1.0.4]: https://github.com/ggsuite/gg_golden/compare/1.0.3...1.0.4
 [1.0.3]: https://github.com/ggsuite/gg_golden/tag/%tag